Perl::Critic clean-up of test suites.

+++ t/60_net_server.t
@@ -5,45 +5,61 @@
 # This is run after the test suite for RPC::XML::Client, so we will use that
 # for the client-side of the tests.
 
+## no critic(RequireCheckedClose)
+## no critic(RequireInterpolationOfMetachars)
+
 use strict;
-use vars qw($dir $srv $pid_file $log_file $port $client $res @keys $meth $list
-            $bucket %seen);
+use warnings;
 use subs qw(start_server find_port);
 
-use Carp qw(croak);
+use Carp qw(carp croak);
 use File::Spec;
+use Module::Load;
 use Test::More;
 
 use RPC::XML::Server;
 use RPC::XML::Client;
 
-eval "use Net::Server";
-# If they do not have Net::Server, quietly skip
-plan skip_all => 'Net::Server not available' if $@;
-# ...or if they are on Windows, skip
-plan skip_all => 'Net::Server tests not reliable on Windows platform'
-    if ($^O eq "MSWin32");
-# otherwise...
-plan tests => 30;
+my ($dir, $srv, $pid_file, $log_file, $port, $client, $res, @keys, $meth, $list,
+    $bucket, %seen);
+
+if ($^O eq 'MSWin32')
+{
+    # Can't run this (reliably) under Windows:
+    plan skip_all => 'Net::Server tests not reliable on Windows platform';
+}
+elsif (! eval { load Net::Server; 1; })
+{
+    # If they do not have Net::Server, quietly skip
+    plan skip_all => 'Net::Server not available';
+}
+else
+{
+    # otherwise...
+    plan tests => 30;
+}
 
 (undef, $dir, undef) = File::Spec->splitpath(File::Spec->rel2abs($0));
 require File::Spec->catfile($dir, 'util.pl');
 
 $pid_file  = File::Spec->catfile($dir, 'net_server.pid');
 $log_file  = File::Spec->catfile($dir, 'net_server.log');
-die "No usable port found between 9000 and 11000, skipping"
-    if (($port = find_port) == -1);
+if (($port = find_port) == -1)
+{
+    croak 'No usable port found between 9000 and 11000, skipping';
+}
 
-unlink $log_file if (-e $log_file);
-unlink $pid_file if (-e $pid_file);
+unlink $log_file, $pid_file;
 
 # All this, and we haven't even created a server object or run a test yet
 
 $srv = RPC::XML::Server->new(no_http => 1);
 # Technically, this is overkill. But if it fails everything else blows up:
 isa_ok($srv, 'RPC::XML::Server');
-croak "Server allocation failed, cannot continue. Message was: $srv"
-    unless (ref $srv);
+if (! ref $srv)
+{
+    croak "Server allocation failed, cannot continue. Message was: $srv";
+}
 
 # All of these parameters are passed to the run() method of
 # Net::Server::MultiType
@@ -63,18 +79,22 @@ $client = RPC::XML::Client->new("http://localhost:$port");
 is($client->simple_request('system.identity'), $srv->product_tokens,
    'system.identity matches $srv->product_tokens');
 
-# At this point, most of this is copied from the first server test suite (40).
-# We do want to verify the full introspection API under Net::Server, though.
+# At this point, most of this is copied from the first server test suite
+# (40_server.t). We do want to verify the full introspection API under
+# Net::Server, though.
 
 $res = $client->simple_request('system.listMethods');
 @keys = $srv->list_methods;
 is(ref($res), 'ARRAY', 'system.listMethods returned ARRAY ref');
 SKIP: {
-    skip 'server response not an ARRAY reference', 2 unless ref($res);
+    if (! ref $res)
+    {
+        skip 'server response not an ARRAY reference', 2;
+    }
 
-    is(scalar(@$res), scalar(@keys),
+    is(scalar(@{$res}), scalar(@keys),
        'system.listMethods returned correct number of names');
-    is(join('', sort @$res), join('', sort @keys),
+    is(join(q{} => sort @{$res}), join(q{} => sort @keys),
        'system.listMethods returned matching set of names');
 }
 
@@ -82,9 +102,12 @@ SKIP: {
 $res = $client->simple_request('system.listMethods', 'method');
 is(ref($res), 'ARRAY', 'system.listMethods returned ARRAY ref');
 SKIP: {
-    skip 'server response not an ARRAY reference', 1 unless ref($res);
+    if (! ref $res)
+    {
+        skip 'server response not an ARRAY reference', 1;
+    }
 
-    is(join(',', sort @$res), 'system.methodHelp,system.methodSignature',
+    is(join(q{,} => sort @{$res}), 'system.methodHelp,system.methodSignature',
        'system.listMethods with pattern returned correct set of names');
 }
 
@@ -92,19 +115,27 @@ SKIP: {
 $res = $client->simple_request('system.listMethods', 'none_will_match');
 is(ref($res), 'ARRAY', 'system.listMethods returned ARRAY ref');
 SKIP: {
-    skip 'server response not an ARRAY reference', 1 unless ref($res);
-    is(scalar(@$res), 0, 'system.listMethods with bad pattern returned none');
+    if (! ref $res)
+    {
+        skip 'server response not an ARRAY reference', 1;
+    }
+
+    is(scalar(@{$res}), 0, 'system.listMethods with bad pattern returned none');
 }
 
 # system.status
 $res = $client->simple_request('system.status');
 @keys = qw(host port name version path date date_int started started_int
            total_requests methods_known);
-is(ref($res), 'HASH', 'system.listMethods returned HASH ref');
+is(ref($res), 'HASH', 'system.status returned HASH ref');
 SKIP: {
-    skip 'server response not a HASH reference', 2 unless ref($res);
-    is(scalar(grep(defined $res->{$_}, @keys)), scalar(@keys),
-       'system.status hashref has correct number of keys');
+    if (! ref $res)
+    {
+        skip 'server response not a HASH reference', 2;
+    }
+
+    my @seen_keys = grep { defined $res->{$_} } @keys;
+    ok(@keys == @seen_keys, 'system.status hash has correct keys');
     is($res->{total_requests}, 5, 'system.status total_request count correct');
 }
 
@@ -115,11 +146,15 @@ is($res, $srv->get_method('system.identity')->{help},
 
 # system.methodHelp with multiple arguments
 $res = $client->simple_request('system.methodHelp',
-                             [ 'system.identity', 'system.status' ]);
+                               [ 'system.identity', 'system.status' ]);
 is(ref($res), 'ARRAY', 'system.methodHelp returned ARRAY ref');
 SKIP: {
-    skip 'server response not an ARRAY reference', 1 unless ref($res);
-    is(join('', @$res),
+    if (! ref $res)
+    {
+        skip 'server response not an ARRAY reference', 1;
+    }
+
+    is(join(q{} => @{$res}),
        $srv->get_method('system.identity')->{help} .
        $srv->get_method('system.status')->{help},
        'system.methodHelp with specific methods returns correctly');
@@ -129,18 +164,28 @@ SKIP: {
 $res = $client->send_request('system.methodHelp', 'system.bad');
 isa_ok($res, 'RPC::XML::fault', 'system.methodHelp (bad arg) response');
 SKIP: {
-    skip 'server response not an RPC::XML data object', 1 unless ref($res);
+    if (! ref $res)
+    {
+        skip 'server response not an RPC::XML data object', 1;
+    }
+
     like($res->string(), qr/Method.*unknown/,
          'system.methodHelp (bad arg) has correct faultString');
 }
 
 # system.methodSignature
 $res = $client->simple_request('system.methodSignature', 'system.methodHelp');
-is(ref($res), 'ARRAY', 'system.methodHelp returned ARRAY ref');
+is(ref($res), 'ARRAY', 'system.methodSignature returned ARRAY ref');
 SKIP: {
-    skip 'server response not an ARRAY reference', 1 unless ref($res);
-    is(join('', sort (map { join(' ', @$_) } @$res)),
-       join('', sort @{ $srv->get_method('system.methodHelp')->{signature} }),
+    if (! ref $res)
+    {
+        skip 'server response not an ARRAY reference', 1;
+    }
+
+    my $return_sig = join q{} => sort map { join q{ } => @{$_} } @{$res};
+    my $method_sig =
+        join q{} => sort @{$srv->get_method('system.methodHelp')->{signature}};
+    is($return_sig, $method_sig,
        'system.methodSignature return value correct');
 }
 
@@ -148,7 +193,11 @@ SKIP: {
 $res = $client->send_request('system.methodSignature', 'system.bad');
 isa_ok($res, 'RPC::XML::fault', 'system.methodSignature (bad arg) response');
 SKIP: {
-    skip 'server response not an RPC::XML data object', 1 unless ref($res);
+    if (! ref $res)
+    {
+        skip 'server response not an RPC::XML data object', 1;
+    }
+
     like($res->string(), qr/Method.*unknown/,
          'system.methodSignature (bad arg) has correct faultString');
 }
@@ -158,26 +207,32 @@ $list = $client->simple_request('system.introspection');
 $bucket = 0;
 %seen = ();
 SKIP: {
-    skip 'system.introspection call did not return ARRAY ref', 1
-        unless (ref($list) eq 'ARRAY');
-
-    for $res (@$list)
+    if (ref($list) ne 'ARRAY')
     {
-        if ($seen{$res->{name}}++)
+        skip 'system.introspection call did not return ARRAY ref', 1;
+    }
+
+    for my $result (@{$list})
+    {
+        if ($seen{$result->{name}}++)
         {
             # If we somehow get the same name twice, that's a point off
             $bucket++;
             next;
         }
 
-        $meth = $srv->get_method($res->{name});
+        $meth = $srv->get_method($result->{name});
         if ($meth)
         {
-            $bucket++ unless
-                (($meth->{help} eq $res->{help}) &&
-                 ($meth->{version} eq $res->{version}) &&
-                 (join('', sort @{ $res->{signature } }) eq
-                  join('', sort @{ $meth->{signature} })));
+            my $result_sig = join q{} => sort @{$result->{signature}};
+            my $method_sig = join q{} => sort @{$meth->{signature}};
+            # A point off unless all three of these match
+            if (($meth->{help} ne $result->{help}) ||
+                    ($meth->{version} ne $result->{version}) ||
+                    ($result_sig ne $method_sig))
+            {
+                $bucket++;
+            }
         }
         else
         {
@@ -193,14 +248,21 @@ $res = $client->simple_request('system.multicall',
                                [ { methodName => 'system.identity' },
                                  { methodName => 'system.listMethods',
                                    params => [ 'intro' ] } ]);
-is(ref($res), 'ARRAY', 'system.methodHelp returned ARRAY ref');
+is(ref($res), 'ARRAY', 'system.multicall returned ARRAY ref');
 SKIP: {
-    skip 'server response not an ARRAY reference', 2 unless ref($res);
+    if (! ref $res)
+    {
+        skip 'server response not an ARRAY reference', 2;
+    }
+
     is($res->[0], $srv->product_tokens,
        'system.multicall, first return value correct');
   SKIP: {
-        skip 'system.multicall return value second index not ARRAY ref', 1
-            unless (ref($res->[1]) eq 'ARRAY');
+        if (ref($res->[1]) ne 'ARRAY')
+        {
+            skip 'system.multicall return value second index not ARRAY ref', 1;
+        }
+
         is(scalar(@{$res->[1]}), 1,
            'system.multicall, second return value correct length');
         is($res->[1]->[0], 'system.introspection',
@@ -214,8 +276,11 @@ $res = $client->send_request('system.multicall',
                                { methodName => 'system.multicall',
                                  params => [ 'intro' ] } ]);
 SKIP: {
-    skip 'system.multicall (recursion) response error, cannot test', 1 unless
-        (ref($res) eq 'RPC::XML::fault');
+    if (ref($res) ne 'RPC::XML::fault')
+    {
+        skip 'system.multicall (recursion) response error, cannot test', 1;
+    }
+
     like($res->string, qr/Recursive/,
          'system.multicall recursion attempt set correct faultString');
 }
@@ -226,8 +291,11 @@ $res = $client->send_request('system.multicall',
                                { methodName => 'system.listMethods',
                                  params => 'intro' } ]);
 SKIP: {
-    skip 'system.multicall (bad data) response error, cannot test', 1 unless
-        (ref($res) eq 'RPC::XML::fault');
+    if (ref($res) ne 'RPC::XML::fault')
+    {
+        skip 'system.multicall (bad data) response error, cannot test', 1;
+    }
+
     like($res->string, qr/value for.*params.*not an array/i,
          'system.multicall bad param array set correct faultString');
 }
@@ -235,21 +303,36 @@ SKIP: {
 # system.status, once more, to check the total_requests value
 $res = $client->simple_request('system.status');
 SKIP: {
-    skip 'system.status response not HASH ref', 1 unless (ref($res) eq 'HASH');
+    if (ref($res) ne 'HASH')
+    {
+        skip 'system.status response not HASH ref', 1;
+    }
+
     is($res->{total_requests}, 19,
        'system.status total_request correct at end of suite');
 }
 
 # Now that we're done, kill the server and exit
-open(my $fh, "< $pid_file");
-chomp(my $pid = <$fh>);
-if ($pid =~ /^(\d+)$/)
+if (open my $fh, '<', $pid_file)
 {
-    kill 'INT', $1;
+    chomp(my $pid = <$fh>);
+    close $fh;
+    if ($pid =~ /^(\d+)$/)
+    {
+        kill 'INT', $1;
+    }
+    else
+    {
+        carp "WARNING: $pid_file appears corrupt, zombie processes may remain!";
+    }
 }
 else
 {
-    warn "WARNING: $pid_file appears corrupt, zombie processes may remain!\n";
+    carp "WARNING: Opening $pid_file failed: $! (zombie processes may remain)";
 }
+
+# If we are exiting cleanly, remove the log file. If we croak'd, we should
+# never reach this line:
+unlink $log_file;
 
 exit;
